+++ CHANGELOG.md
@@ -2,10 +2,28 @@
 
 ## Unreleased
 
-## Added
+### Added
 
+* `/v3/manage_offer_requests` endpoint
+* `/v3/create_payment_requests` endpoint
+* `CreateManageOfferRequestOp` ingestion
+* `CreatePaymentRequestOp` ingestion
+* `CreatePaymentRequestOp` ingestion
+* `PaymentRequest` ingestion
+* `ManageOfferRequest` ingestion
 * `RemoveAssetOp` ingestion
 
+
+### Fixed
+
+* panic on `v3/balances`
+
+## 3.5.2-x.0
+
+### Fixed
+
+* Key for atomic swap ask quote asset relation
+* Panic on `ManagePoll` ingestion
 
 ## 3.5.1